+++ src/HandlerResolver/InjectorExecutableResolver.php
@@ -13,7 +13,7 @@ declare(strict_types = 1);
 namespace Cspray\Labrador\Http\HandlerResolver;
 
 use Auryn\Injector;
-use Symfony\Component\HttpFoundation\Request;
+use Psr\Http\Message\ServerRequestInterface;
 
 class InjectorExecutableResolver implements HandlerResolver {
 
@@ -28,11 +28,12 @@ class InjectorExecutableResolver implements HandlerResolver {
     /**
      * If the implementation cannot turn $handler into a callable type return false.
      *
+     * @param ServerRequestInterface $request
      * @param mixed $handler
      * @return callable|false
      * @throws \Cspray\Labrador\Http\Exception\InvalidHandlerException
      */
-    public function resolve(Request $request, $handler) {
+    public function resolve(ServerRequestInterface $request, $handler) {
         $cb = $this->resolver->resolve($request, $handler);
         if ($cb) {
             $injector = $this->injector;